refactor(sdf): move writer to temporary namespace
