+++ mupt/interfaces/rdkit/__init__.py
@@ -31,7 +31,7 @@ from .strategies import (
     RDKitExportStrategy,
     AllAtomRDKitExportStrategy,
 )
-from ...muptio.sdf import write_primitive_to_mupt_sdf, write_primitive_to_sdf
+from ...temporary.sdf import write_primitive_to_mupt_sdf, write_primitive_to_sdf
 from .depiction import (
     set_rdkdraw_size,
     show_substruct_highlights,
