+++ mupt/tests/temporary/__init__.py
@@ -1,4 +1,4 @@
-"""Tests for MuPT file I/O helpers."""
+"""Tests for temporary MuPT helpers."""
 
 __author__ = "Joseph R. Laforet Jr."
 __email__ = "jola3134@colorado.edu"
